cherry pick Add tests for get_item_id
