+++ tests/test_graphics_mem_safety.py
@@ -2,27 +2,8 @@
 import pytest
 import imgui
 
-@pytest.fixture
-def context():
-    return imgui.create_context()
-
-@pytest.fixture
-def io():
-    # setup io
-    io = imgui.get_io()
-    io.delta_time = 1.0 / 60.0
-    io.display_size = 300, 300
-
-    # setup default font
-    io.fonts.get_tex_data_as_rgba32()
-    io.fonts.add_font_default()
-    io.fonts.texture_id = 42  # set any texture ID to avoid segfaults
-
-    return io
-
 @pytest.mark.flaky(reruns=10, reruns_delay=5)
 def test_texture_id_int_reference(context, io):
-
     # See issue #248 (https://github.com/pyimgui/pyimgui/issues/248)
 
     texture_id = 0
@@ -50,9 +31,9 @@ def test_texture_id_int_reference(context, io):
                 print(' - type texture id:', type(command.texture_id))
                 assert type(command.texture_id) is int
 
-def test_texture_id_keep_type(context, io):
 
-    texture_id = { 'dummy':42 }
+def test_texture_id_keep_type(context, io):
+    texture_id = {'dummy': 42}
 
     imgui.new_frame()
     imgui.image(texture_id, 640, 480)
@@ -65,4 +46,4 @@ def test_texture_id_keep_type(context, io):
             if command.texture_id == io.fonts.texture_id:
                 continue
 
-            assert type(command.texture_id) == type(texture_id)    
+            assert type(command.texture_id) == type(texture_id)
